refactor($step4): MVC 패턴 기반으로 패키지 분리
