+++ src/main/java/racing/domain/movestrategies/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racing.movestrategies;
+package racing.domain.movestrategies;
 
 @FunctionalInterface
 public interface MoveStrategy {
