+++ src/test/java/racing/domain/RacingGameTest.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
